refactor(firmware): clear every clang-tidy finding
Nothing here changes what the firmware does. What it changes is what the
tree can say about itself: `make tidy` now reports zero findings, so the
config can fail the build on the next one instead of printing it.

Three were worth the trip on their own. An AdslPacket walked 24 bytes off
&Version, a single uint8_t, which is undefined behaviour the analyser was
right about: the wire bytes are a named Data[kDataBytes] member of the
union that was already there. answer_read() divided by a sector count a
flash part is allowed to report as zero. load_scenario() parsed a
truncated read as a whole scenario, because nothing asked ferror().

The rest is the C++17 the tree was already writing longhand: std::min,
std::max and std::clamp where a pair of ifs was a clamp, std::any_of for
a search loop, explicit casts where a multiplication was widened after
the fact, static and const on members that touch no state, and a
declaration whose parameter names match its definition.

+++ firmware/hardware/platform/host/flash_region.h
@@ -24,7 +24,8 @@ class FlashRegion : public ports::FlashRegion {
     static constexpr uint32_t kSectorCount = 330;
 
     explicit FlashRegion(uint32_t sector_count = kSectorCount)
-        : sector_count_(sector_count), bytes_(sector_count * kSectorBytes, 0xFF) {}
+        : sector_count_(sector_count),
+          bytes_(static_cast<size_t>(sector_count) * kSectorBytes, 0xFF) {}
 
     bool ready() const override { return present_; }
     void set_present(bool on) { present_ = on; }
@@ -61,7 +62,7 @@ class FlashRegion : public ports::FlashRegion {
         if (!present_) return Status::Down;
         if (index >= sector_count_) return Status::OutOfRange;
         if (dead_) return Status::Down;
-        std::memset(bytes_.data() + index * kSectorBytes, 0xFF, kSectorBytes);
+        std::memset(bytes_.data() + static_cast<size_t>(index) * kSectorBytes, 0xFF, kSectorBytes);
         erases++;
         return Status::Ok;
     }
